requirements.txt
aiodns==3.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
APScheduler==3.11.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
backports.asyncio.runner==1.2.0
beautifulsoup4==4.13.4
ccxt==4.5.17
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
coincurve==20.0.0
contourpy==1.3.2
cryptography==46.0.3
cssselect==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.1
distlib==0.4.0
exceptiongroup==1.3.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.4
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
Js2Py==0.74
kiwisolver==1.4.8
lxml==5.4.0
m3u8==6.0.0
matplotlib==3.10.3
mistletoe==1.4.0
mplfinance==0.12.10b0
multidict==6.5.0
nested-lookup==0.2.25
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parsel==1.10.0
pillow==11.2.1
platformdirs==4.3.8
playwright==1.51.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pyaes==1.6.1
pyasn1==0.6.1
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pyee==12.1.1
Pygments==2.19.1
pyjsparser==2.7.1
pymexc==1.2.12
pyotp==2.9.0
pyparsing==3.2.3
pyTelegramBotAPI==4.27.0
pytest==8.4.1
pytest-asyncio==1.2.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python-telegram-bot==22.0
pytz==2024.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.28.2
rsa==4.9.1
schedule==1.1.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
ta==0.11.0
telegramify-markdown==0.5.1
Telethon==1.40.0
text-unidecode==1.3
tomli==2.2.1
twikit==2.3.1
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
virtualenv==20.33.1
w3lib==2.3.1
websocket-client==1.8.0
websockets==15.0.1
webvtt-py==0.5.1
wsaccel==0.6.7
yarl==1.20.1
